+++ modules/aprs/handlers/yb-tokens/yb-apr-handlers/sources/default-apr-handler.ts
@@ -54,8 +54,7 @@ export class DefaultAprHandler implements YbAprHandler {
                 },
             };
         } catch (error) {
-            console.error(`Failed to fetch APRs in url ${this.url}:`, error);
-            return {};
+            throw Error(`Failed to fetch APRs in url ${this.url}: ${(error as Error).message}`);
         }
     }